requirements.txt
torch
torchvision
typer
wandb
matplotlib
scikit-learn
hydra-core
omegaconf